+++ protocols/gossipsub/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.47.1
-- Implement backpressure by diferentiating between priority and non priority messages.
+- Implement backpressure by differentiating between priority and non priority messages.
   Drop `Publish` and `Forward` messages when the queue becomes full.
   See [PR 4914](https://github.com/libp2p/rust-libp2p/pull/4914)