+++ protocols/gossipsub/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.47.0
 
+- Attempt to publish to at least mesh_n peers when flood publish is disabled.
+  See [PR XXXX](https://github.com/libp2p/rust-libp2p/pull/XXXX).
+
+## 0.47.0
+
 <!-- Update to libp2p-swarm v0.45.0 -->
 - Add ConnectionError to FromSwarm::ConnectionClosed.
   See [PR 5485](https://github.com/libp2p/rust-libp2p/pull/5485).